% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{assign}
\alias{assign}
\title{Taxonomic assignment of metabarcoding sequencing reads.}
\source{
(1) Somervuo et al. Methods in Ecology & Evolution 2017: 
\url{https://doi.org/10.1111/2041-210X.12721}
}
\usage{
assign(in_dir = NA, out_dir = NA, marker = "16S", model_dir = NA, sfb_dir = NA)
}
\arguments{
\item{in_dir}{Folder containing demultiplexed and dereplicated sequencing reads.}

\item{out_dir}{Custom name of the created output folder that is created. Default is "AssignOut_" followed by marker and current date and time (e.g. "AssignOut_16S_20221219_110213")}

\item{marker}{String defineing the genetic marker used for metabarcoding. Must be can be one or several of "16S", "12S" or "CytB" (e.g. marker = c("12S", "16S"). Default is marker = "16S".}

\item{model_dir}{Path to a directory containing PROTAX models and clean databases for all loci. Default uses the latest model from '/home/bioadmin/Protax/models'.}

\item{sfb_dir}{Path to the screenforbio-mbc directory and it's subdirectory '/protaxscripts' (screenforbio pipeline of Alex Crampton-Platt).}
}
\description{
assign
}
\details{
Function deploys the bash scripts for PROTAX-assignments of metabarcoding sequencing reads from the screenforbio pipeline of Alex Crampton-Platt. It can be used to call either weighted or unweighted PROTAX models ((1) Somervuo et al. 2017). PROTAX models must have been pre-trained using \code{train_protax}.
}
\examples{
Simple example to analyse 16S sequencing reads with using th default settings:
assign(in_dir = "/SequencingData/221114_M01108_0115/seqRun001/data/derep")

}
