% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignReads.R
\name{assignReads}
\alias{assignReads}
\title{Taxonomic assignment of metabarcoding sequencing reads.}
\source{
(1) Somervuo et al. Methods in Ecology & Evolution 2017: 
\url{https://doi.org/10.1111/2041-210X.12721}

#' @source
(2) \code{last}: 
\url{https://gitlab.com/mcfrith/last}

#' (3) \code{vsearch}: 
\url{https://github.com/torognes/vsearch}
}
\usage{
assignReads(
  in_dir = NA,
  out_dir = NA,
  marker = "16S",
  model_dir = NA,
  num_cores = 20
)
}
\arguments{
\item{in_dir}{Folder containing de-multiplexed and de-replicated sequencing reads.}

\item{out_dir}{Custom name of the created output folder that is created. Default is "AssignOut_" followed by marker and current date and time (e.g. "AssignOut_16S_20221219_110213")}

\item{marker}{String defining the genetic marker used for metabarcoding. Must be can be one or several of "16S", "12S" or "CytB" (e.g. marker = c("12S", "16S"). Default is marker = "16S".}

\item{model_dir}{Path to a directory containing PROTAX models and clean databases for all loci. Default uses the latest model from '/home/bioadmin/Protax/models'.}
}
\description{
assignReads
}
\details{
Function deploys the bash scripts for PROTAX-assignments of metabarcoding sequencing reads from the screenforbio pipeline of Alex Crampton-Platt. It can be used to call either weighted or unweighted PROTAX models ((1) Somervuo et al. 2017). PROTAX models must have been pre-trained using \code{train_protax}. The bash script uses \code{last} (2) and \code{vsearch} (3), which must be installed on the computer running the analysis and their superior directory containing its their executable files must be added to the PATH variable of the environment (e.g. use base::Sys.setenv() to add them in R).
}
