% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{demultiplex}
\alias{demultiplex}
\title{Demultiplex double indexed Illumina sequencing reads.}
\source{
(1) \code{bcl2fastq}: 
\url{https://emea.support.illumina.com/sequencing/sequencing_software/bcl2fastq-conversion-software.html}

(2) \code{cutadapt}:
\url{https://doi.org/10.14806/ej.17.1.200}

(3) \code{adapterremoval}:  
\url{https://bmcresnotes.biomedcentral.com/articles/10.1186/s13104-016-1900-2}

(4) \code{usearch}:
\url{https://www.drive5.com/usearch/}

(5) \code{vsearch}: 
\url{https://github.com/torognes/vsearch}

(6) \code{flash}:
\url{https://}
}
\usage{
demultiplex(
  seq_dir = NA,
  seq_run = NA,
  out_dir = NA,
  marker = "16S",
  cores = 24,
  cycles = NA,
  paired_end = T,
  merge = NA,
  demulti_b = T,
  demulti_s = T
)
}
\arguments{
\item{seq_dir}{Run folder of an Illumina sequencing run.}

\item{seq_run}{Optional sequencing run name (e.g. seq_run = "LeechSeq001"). if not provided it takes the 'Experiment Name' from the SampleSheet.}

\item{out_dir}{Optional custom name of the created output folder that is created in the run folder. Default is "DerepOut_" followed by current date and time (e.g. "DerepOut_20221219_110204")}

\item{marker}{Optinal character giving the sequenced genetic markers. By default it is only "16S", but "12S" and "cytB" are also possible, e.g. 'marker = c("16S", "12S", "cytB")' would use all three options.}

\item{cycles}{Optional integer giving the number of sequencing cycles, per default read from the SampleSheet of the run.}

\item{paired_end}{Logical parameter ste to 'paired_end = TRUE' by default, indicating if it was a paired end  or single read sequencing run.}

\item{merge}{Optional parameter defining the programme used for merging paired end reads. By default \code{usearch} will be used but alternatively it can be set to \code{vsearch} or \code{flash}, e.g. 'merge = vsearch'.}

\item{demulti_b}{Logical parameter defining if PCR batches shall be demultiplexed. By default it is set to 'demulti_b = TRUE'. If set to 'FALSE' only fastq files will be generated.}

\item{demulti_s}{Logical parameter defining if a second demultiplexing step shall be performed. By default it is set to 'demulti_b = TRUE'. If set to 'FALSE' the reads will only be demultiplexed to PCR batch level and not to sample level.}
}
\description{
demultiplex
}
\details{
Function deploys several steps previously combined in the read_preprocessing.sh script from the screenforbio pipeline of Alex Crampton-Platt. It uses \code{bcl2fastq} (1), \code{cudadapt} (2), \code{adapterremoval} (3) and, optional \code{usearch}(4), \code{vsearch} (5) or \code{flash} (6). All programmes must be installed on the computer running the analysis and their superior directory where their executable files are saved must be added to the PATH variable of the environment (e.g. use base::Sys.setenv() to add them in R). Do not use the project column in the SampleSheet, otherwise \code{bc2fastq} creates subfolders that cannot be handled by the script.
}
