% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_queries.r
\name{get_reads}
\alias{get_reads}
\title{Query sequencing reads from database and convert them to fasta file}
\usage{
get_reads(
  sample_id = NA,
  tax_level = "species",
  tax_query = NA,
  read_freq = 100,
  dir = getwd(),
  file_out = NA,
  db_name = "biodiv",
  db_user = "r_select",
  db_pwd = "rs#izw22",
  db_host = "192.168.2.71",
  db_port = "5433"
)
}
\arguments{
\item{sample_id}{Sample IDs as they are documented in the database (mandatory). String or vector of strings of complete or incomplete IDs as wildcards are added (e.g. c("VBDC", "VBF") or "VBDC_12_1_").}

\item{tax_level}{Taxonomic level at which you query (mandatory). If you want to query at other taxonomic levels than the default "species" you can set it to either "genus" or "family".}

\item{tax_query}{String or vector of strings of Latin taxon names that you wish to query (mandatory). Strings can be complete or incomplete taxonomic names (e.g. "Canis lupus" or c("Canis", "Felis")). Keep in mind that tax_query must fit the tax_level your perform your query on.}

\item{read_freq}{Numeric threshold for the minimum frequency of a read (optional). Default is 100.}

\item{dir}{Custom output directory (optional), default is the current working directory (getwd()).}

\item{file_out}{Custom name of output file (optional). Output file is in fasta format ("*.fas") providing sequence ID, taxon and read frequency in the sequence header. Default creates a file name starting with "ReadsOut", followed by tax_query and date (e.g. "ReadsOut_Canis_20221202.fas").}

\item{db_name}{String giving the database name (optional). Default is the biodiv database.}

\item{db_user}{String giving the name of the database user (optional). Default user has only SELECT privileges.}

\item{db_pwd}{String giving the password of the database user (optional). Default user has only SELECT privileges.}

\item{db_host}{String giving the internal IP address of the host server of the biodiv database. If you wish to connect to another database you have to adjust db_name and db_host accordingly. To connect to the PostgreSQL database from outside the IZW domain it would need ssh connection to the database host and port forwarding to 5433. The db_host must be set to "localhost" then.}

\item{db_port}{String giving the port the Postgres server listens to. Default is set to "5433", the port bidov database listens to.}
}
\description{
The get_reads function queries read sequences assigned to a certain taxon and writes them to fasta file. The query is based in sample_id, assigned taxon and taxonomic level. To connect to the PostgreSQL database from outside the IZW domain it would need ssh connection to the database host and port forwarding to 5433. The db_host must be set to "localhost" then.
}
