% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_queries.r
\name{get_refs}
\alias{get_refs}
\title{Query reference sequences from database and convert them to fasta file}
\usage{
get_refs(
  tax_query = NA,
  gen_marker = "16S",
  dir = getwd(),
  file_out = NA,
  db_name = "biodiv",
  db_user = "R_select",
  db_pwd = "rs#izw22",
  db_host = "192.168.2.71",
  db_port = "5433"
)
}
\arguments{
\item{tax_query}{String or vector of strings of Latin taxon names that you wish to query (mandatory). Strings can be complete or incomplete taxonomic names (e.g. "Canis lupus" or c("Canis", "Felis")).}

\item{gen_marker}{String defining the genetic marker, could be either "16S", "12S" or "CytB". Default is set to "16S".}

\item{dir}{Custom output directory (optional), default is the current working directory (getwd()).}

\item{file_out}{Custom name of output file (optional). Output file is in fasta format ("*.fas") providing taxon, its reference ID in the sequence header. Default creates a file name starting with "RefsOut", followed by taxon, genetic marker and date (e.g. "RefsOut_Canis_16S_20221202.fas").}

\item{db_name}{String giving the database name (optional). Default is the biodiv database.}

\item{db_user}{String giving the name of the database user (optional). Default user has only SELECT privileges.}

\item{db_pwd}{String giving the password of the database user (optional). Default user has only SELECT privileges.}

\item{db_host}{String giving the internal IP address of the host server of the biodiv database. If you wish to connect to another database you have to adjust db_name and db_host accordingly. To connect to the PostgreSQL database from outside the IZW domain it would need ssh connection to the database host and port forwarding to 5433. The db_host must be set to "localhost" then.}

\item{db_port}{String giving the port the Postgres server listens to. Default is set to "5433", the port bidov database listens to.}
}
\description{
get_refs
}
\details{
The get_refs function, queries reference sequences and writes them to fasta file.
The query is based in taxon and genetic marker. To connect to the PostgreSQL database from outside the IZW domain it would need ssh connection to the database host and port forwarding to 5433. The db_host must be set to "localhost" then.
}
\examples{
Simple query with default settings: 
get_refs(tax_query = c("Canis", "Felis"))
}
