% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_tab.R
\name{rep_tab}
\alias{rep_tab}
\title{Aggregate assignment results per sample and replicate ID.}
\usage{
rep_tab(
  sample_id = NA,
  dir = getwd(),
  file_out = NA,
  db_name = "biodiv",
  db_user = "r_select",
  db_pwd = "rs#izw22",
  db_host = "192.168.2.71",
  db_port = "5433"
)
}
\arguments{
\item{sample_id}{String defining a sample ID or study site (e.g. VBDC = BiDoup National Park). For a list of current study sites and their abbreviations use study_sites(). Strings can be complete or incomplete sample IDs (e.g. "VBDC" or c("VBDC_068_1_1", "VBDC_062_1_2")).}

\item{dir}{Custom output directory (optional), default is the current working directory (getwd()).}

\item{file_out}{Custom name of output file (optional). Output file is a tab delimited text file ("*.txt"). Default creates a file name starting with "RepAssignResOut", followed by sample_id and date (e.g. "RepAssignResOut_VDBC_068_1_1_rep1_20221207_105429.txt").}

\item{db_name}{String giving the database name (optional). Default is the biodiv database.}

\item{db_user}{String giving the name of the database user (optional). Default user has only SELECT privileges.}

\item{db_pwd}{String giving the password of the database user (optional). Default user has only SELECT privileges.}

\item{db_host}{String giving the internal IP address of the host server of the biodiv database. If you wish to connect to another database you have to adjust db_name and db_host accordingly. To connect to the PostgreSQL database from outside the IZW domain it would need ssh connection to the database host and port forwarding to 5433. The db_host must be set to "localhost" then.}

\item{db_port}{String giving the port the Postgres server listens to. Default is set to "5433", the port biodiv database listens to.}
}
\description{
rep_tab
}
\details{
Lists the assignments for the different taxonomic levels and calculated the mean log probability for those assignments made by PROTAX. Also lists the reference with the highest similarity and the average similarity score. To connect to the PostgreSQL database from outside the IZW domain it would need ssh connection to the database host and port forwarding to 5433. The db_host must be set to "localhost" then.
}
