% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_tab.R
\name{res_tab}
\alias{res_tab}
\title{Aggregate assignment results per sample and replicate ID.}
\usage{
res_tab(
  sample_id = NA,
  rep_id = NA,
  dir = getwd(),
  file_out = NA,
  db_name = "biodiv",
  db_user = "r_select",
  db_pwd = "rs#izw22",
  db_host = "192.168.2.71",
  db_port = "5433"
)
}
\arguments{
\item{sample_id}{String defining a sample ID (optional). Strings can be complete or incomplete sample IDs (e.g. "VDBC" or c("VBDC068_1_1", "VBDC_062_1_2")). Either sample_id or rep_id (or both) must be provided.}

\item{rep_id}{String defining a replicate ID (optional). Strings can be complete or incomplete replicate IDs (e.g. "rep1" or c("rep1", "rep2")). Either sample_id or rep_id (or both) must be provided.}

\item{dir}{Custom output directory (optional), default is the current working directory (getwd()).}

\item{file_out}{Custom name of output file (optional). Output file is a tab delimited text file ("*.txt"). Default creates a file name starting with "AssignResOut", followed by sample_id and rep_id (if provided) and date (e.g. "AssignResOut_20221207_105429.txt" or "AssignResOut_VBDC_068_1_1_rep1_20221207_105429.txt").}

\item{db_name}{String giving the database name (optional). Default is the biodiv database.}

\item{db_user}{String giving the name of the database user (optional). Default user has only SELECT privileges.}

\item{db_pwd}{String giving the password of the database user (optional). Default user has only SELECT privileges.}

\item{db_host}{String giving the internal IP address of the host server of the biodiv database. If you wish to connect to another database you have to adjust db_name and db_host accordingly. To connect to the PostgreSQL database from outside the IZW domain it would need ssh connection to the database host and port forwarding to 5433. The db_host must be set to "localhost" then.}

\item{db_port}{String giving the port the Postgres server listens to. Default is set to "5433", the port biodiv database listens to.}
}
\description{
res_tab
}
\details{
Lists the assignments for the different taxonomic levels and calculated the mean log probability for those assignments made by PROTAX. Also lists the reference with the highest similarity and the average similarity score. To connect to the PostgreSQL database from outside the IZW domain it would need ssh connection to the database host and port forwarding to 5433. The db_host must be set to "localhost" then.
}
