% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_protax.R
\name{train_protax}
\alias{train_protax}
\title{train Protax models for taxonomic asignment of metabarcoding reads.}
\usage{
train_protax(
  in_dir = NA,
  splist = NA,
  out_dir = NA,
  marker = "16S",
  weight = 0.9,
  taxon = "Tetrapoda",
  cores = 1
)
}
\arguments{
\item{in_dir}{Input directory containing the reference taxonomy and reference databases containing the reference sequences. The latter must be fasta files and follow the name schema 'taxon.database.date.marker.fa' (e.g. 'Tetrapoda.database.20220315.12S.fa').}

\item{splist}{splist is a list of expected species to use in weighing in the format 'Genus,species' (e.g. 'Homo,sapiens').}

\item{out_dir}{Optional output directory. If not provided an output folder including date and time is created (e.g. 'TrainProtax_20240722_170211')}

\item{marker}{Optinal character giving the sequenced genetic markers. By default it is only "16S", but "12S" and "cytB" are also possible, e.g. 'marker = c("16S", "12S", "cytB")' would use all three options.}

\item{weight}{Optional numeric variable giving the weight for the expected species. Must be between 0 and 1.}

\item{taxon}{Reference taxonomy used for Protax training. File must follow the name schema '*.protax_taxonomy.date.txt' (e.g. 'Tetrapoda.protax_taxonomy.20220318.txt').}

\item{cores}{Optional parameter that can be used to parallelize 'lastal' in step 3, which speeds-up processing. Default is set 'cores = 1', which means a single thread is running.}
}
\description{
train_protax
}
\details{
Function deploys the train_weighted_protax.sh script from the screenforbio pipeline of Alex Crampton-Platt. It will take the taxon from the protax taxonomy file name and assumes curated database FASTA files are in current directory and labelled with format '*.final_database.marker.fa'. The bash script uses \code{last}, which must be installed on the computer running the analysis and their superior directory containing its their executable files must be added to the PATH variable of the environment (e.g. use base::Sys.setenv() to add them in R).
}
